refactor: separate defuse-fees crate
